Fix travis build on HHVM
Travis build fails because no HHVM version is specified, travis then tries
to build on a pretty old 3.3 version. Unfortunately, php-cs-fixer conflicts
with HHVM prior to version 3.9.

In this commit, we force an Ubuntu trusty in order to be able to choose a more
recent HHVM version, and set it to 3.12.

see https://docs.travis-ci.com/user/languages/php/#HHVM-versions for further details